+++ main.py
@@ -1,7 +1,7 @@
 # Package: BulkDNS
 # Module: main
 # Author: Michal Selma <michal@selma.cc>
-# Rev: 2023-12-10
+# Rev: 2023-12-15
 
 import os
 import sys
@@ -9,20 +9,22 @@ import configparser
 import logging
 
 from init import init_core
+from core import proc_core
 
 from common import logger
 
 log = logging.getLogger('main')
 
 if __name__ == '__main__':
-    # init logging configuration for root logger
-    logger.configure(log)
-
     # Read configuration
     config = configparser.ConfigParser()
     if not config.read(os.path.realpath('./config.cfg')):
-        log.critical(f'Configuration file not found.')
+        print(f'Configuration file not found.')
         sys.exit(0)
+
+    # When configuration is loaded modify logging configuration if needed
+    log_level = config['DEFAULT']['log_level']
+    logger.configure(log, log_level)
 
     log.info('1 - System initialization')
     log.info('2 - Domains check')
@@ -32,7 +34,7 @@ if __name__ == '__main__':
     if user_option == '1':
         init_core.run(config)
     elif user_option == '2':
-        log.info('Not implemented yet...')
+        proc_core.run(config)
     elif user_option == '3':
         log.info('Not implemented yet...')
     else: